+++ refena/lib/refena.dart
@@ -15,6 +15,8 @@ export 'package:refena/src/container.dart' show PlatformHint, RefenaContainer;
 export 'package:refena/src/notifier/base_notifier.dart'
     show
         NotifyStrategy,
+        RebuildableNotifier,
+        FutureProviderNotifier,
         ViewProviderNotifier,
         ViewFamilyProviderNotifier,
         ReduxNotifier,
@@ -30,7 +32,6 @@ export 'package:refena/src/notifier/rebuildable.dart';
 export 'package:refena/src/notifier/types/async_notifier.dart';
 export 'package:refena/src/notifier/types/change_notifier.dart';
 export 'package:refena/src/notifier/types/future_family_provider_notifier.dart';
-export 'package:refena/src/notifier/types/future_provider_notifier.dart';
 export 'package:refena/src/notifier/types/immutable_notifier.dart';
 export 'package:refena/src/notifier/types/notifier.dart';
 export 'package:refena/src/notifier/types/pure_notifier.dart';