+++ refena/lib/src/container.dart
@@ -4,15 +4,14 @@ import 'dart:collection';
 import 'package:meta/meta.dart';
 import 'package:refena/src/accessor.dart';
 import 'package:refena/src/action/dispatcher.dart';
+import 'package:refena/src/async_value.dart';
 import 'package:refena/src/notifier/base_notifier.dart';
 import 'package:refena/src/notifier/family_notifier.dart';
 import 'package:refena/src/notifier/notifier_event.dart';
-import 'package:refena/src/notifier/types/async_notifier.dart';
 import 'package:refena/src/observer/event.dart';
 import 'package:refena/src/observer/observer.dart';
 import 'package:refena/src/provider/base_provider.dart';
 import 'package:refena/src/provider/override.dart';
-import 'package:refena/src/provider/types/async_notifier_provider.dart';
 import 'package:refena/src/provider/types/notifier_provider.dart';
 import 'package:refena/src/provider/types/redux_provider.dart';
 import 'package:refena/src/provider/watchable.dart';
@@ -317,8 +316,8 @@ class RefenaContainer implements Ref, LabeledReference {
   }
 
   @override
-  Future<T> future<N extends AsyncNotifier<T>, T>(
-    AsyncNotifierProvider<N, T> provider,
+  Future<T> future<N extends BaseAsyncNotifier<T>, T>(
+    BaseProvider<N, AsyncValue<T>> provider,
   ) {
     // ignore: invalid_use_of_protected_member
     return _getState(provider).future;