+++ refena/lib/src/notifier/types/view_provider_notifier.dart
@@ -1,6 +1,7 @@
 part of '../base_notifier.dart';
 
 final class ViewProviderNotifier<T> extends BaseSyncNotifier<T>
+    with RebuildableNotifier
     implements Rebuildable {
   ViewProviderNotifier(
     this._builder, {
@@ -8,10 +9,8 @@ final class ViewProviderNotifier<T> extends BaseSyncNotifier<T>
     super.debugLabel,
   }) : _describeState = describeState;
 
-  late final WatchableRef _watchableRef;
   final T Function(WatchableRef) _builder;
   final String Function(T state)? _describeState;
-  final _rebuildController = BatchedStreamController<AbstractChangeEvent>();
 
   @override
   T init() {
@@ -19,77 +18,11 @@ final class ViewProviderNotifier<T> extends BaseSyncNotifier<T>
       // rebuild notifier state
       _setStateAsRebuild(
         this,
-        _build(),
+        _callAndTrackNotifiers(_builder),
         event,
       );
     });
-    return _build();
-  }
-
-  T _build() {
-    final oldDependencies = {...dependencies};
-    dependencies.clear();
-
-    final nextState = (_watchableRef as WatchableRefImpl).trackNotifier(
-      onAccess: (notifier) {
-        final added = dependencies.add(notifier);
-        if (!added) {
-          printAlreadyWatchedWarning(
-            rebuildable: this,
-            notifier: notifier,
-          );
-        }
-        notifier.dependents.add(this);
-      },
-      run: () => _builder(_watchableRef),
-    );
-
-    final removedDependencies = oldDependencies.difference(dependencies);
-    for (final removedDependency in removedDependencies) {
-      // remove from dependency graph
-      removedDependency.dependents.remove(this);
-
-      // remove listener to avoid future rebuilds
-      removedDependency._listeners.removeListener(this);
-    }
-
-    return nextState;
-  }
-
-  @internal
-  @override
-  void internalSetup(
-    ProxyRef ref,
-    BaseProvider<BaseNotifier<T>, T>? provider,
-  ) {
-    _watchableRef = WatchableRefImpl(
-      container: ref.container,
-      rebuildable: this,
-    );
-
-    super.internalSetup(ref, provider);
-  }
-
-  @override
-  void rebuild(ChangeEvent? changeEvent, RebuildEvent? rebuildEvent) {
-    assert(
-      changeEvent == null || rebuildEvent == null,
-      'Cannot have both changeEvent and rebuildEvent',
-    );
-
-    if (changeEvent != null) {
-      _rebuildController.schedule(changeEvent);
-    } else if (rebuildEvent != null) {
-      _rebuildController.schedule(rebuildEvent);
-    } else {
-      _rebuildController.schedule(null);
-    }
-  }
-
-  @override
-  @nonVirtual
-  void dispose() {
-    _rebuildController.dispose();
+    return _callAndTrackNotifiers(_builder);
   }
 
   @override
@@ -99,13 +32,4 @@ final class ViewProviderNotifier<T> extends BaseSyncNotifier<T>
     }
     return _describeState!(state);
   }
-
-  @override
-  void onDisposeWidget() {}
-
-  @override
-  void notifyListenerTarget(BaseNotifier notifier) {}
-
-  @override
-  bool get isWidget => false;
 }