+++ refena/lib/src/provider/types/future_provider.dart
@@ -1,8 +1,11 @@
+import 'package:meta/meta.dart';
 import 'package:refena/src/async_value.dart';
-import 'package:refena/src/notifier/types/future_provider_notifier.dart';
+import 'package:refena/src/notifier/base_notifier.dart';
+import 'package:refena/src/provider/base_provider.dart';
 import 'package:refena/src/provider/override.dart';
 import 'package:refena/src/provider/types/async_notifier_provider.dart';
 import 'package:refena/src/provider/types/future_family_provider.dart';
+import 'package:refena/src/proxy_ref.dart';
 import 'package:refena/src/ref.dart';
 
 /// A [FutureProvider] is a custom implementation of a [AsyncNotifierProvider]
@@ -20,24 +23,30 @@ import 'package:refena/src/ref.dart';
 /// - fetch static data from an API (that does not change)
 /// - fetch device information (that does not change)
 class FutureProvider<T>
-    extends AsyncNotifierProvider<FutureProviderNotifier<T>, T> {
-  final String Function(AsyncValue<T> state)? _describeState;
-
+    extends BaseWatchableProvider<FutureProviderNotifier<T>, AsyncValue<T>>
+    with ProviderSelectMixin<FutureProviderNotifier<T>, AsyncValue<T>> {
   FutureProvider(
-    Future<T> Function(Ref ref) builder, {
+    this._builder, {
     super.onChanged,
     String Function(AsyncValue<T> state)? describeState,
     String? debugLabel,
     super.debugVisibleInGraph = true,
   })  : _describeState = describeState,
         super(
-          (ref) => FutureProviderNotifier<T>(
-            builder(ref),
-            describeState: describeState,
-            debugLabel: debugLabel ?? 'FutureProvider<$T>',
-          ),
           debugLabel: debugLabel ?? 'FutureProvider<$T>',
         );
+
+  final Future<T> Function(WatchableRef ref) _builder;
+  final String Function(AsyncValue<T> state)? _describeState;
+
+  @internal
+  @override
+  FutureProviderNotifier<T> createState(ProxyRef ref) {
+    return _createState(
+      ref: ref,
+      builder: _builder,
+    );
+  }
 
   /// Overrides the future.
   ///
@@ -48,10 +57,21 @@ class FutureProvider<T>
     return ProviderOverride<FutureProviderNotifier<T>, AsyncValue<T>>(
       provider: this,
       createState: (ref) => FutureProviderNotifier<T>(
-        builder(ref),
+        builder,
         describeState: _describeState,
         debugLabel: customDebugLabel ?? runtimeType.toString(),
       ),
+    );
+  }
+
+  FutureProviderNotifier<T> _createState({
+    required ProxyRef ref,
+    required Future<T> Function(WatchableRef ref) builder,
+  }) {
+    return FutureProviderNotifier<T>(
+      builder,
+      describeState: _describeState,
+      debugLabel: customDebugLabel ?? runtimeType.toString(),
     );
   }