+++ refena/test/integration/provider/future_provider_test.dart
@@ -4,14 +4,15 @@ import 'package:test/test.dart';
 import '../../util/skip_microtasks.dart';
 
 void main() {
-  test('Should read the value', () async {
-    final provider = FutureProvider((ref) => Future.value(123));
+  test('Should provide an async value', () async {
     final observer = RefenaHistoryObserver.all();
     final ref = RefenaContainer(
       observers: [observer],
     );
 
+    final provider = FutureProvider((ref) => Future.value(123));
     expect(ref.read(provider), AsyncValue<int>.loading());
+
     expect(await ref.future(provider), 123);
     expect(
       ref.read(provider),
@@ -40,92 +41,190 @@ void main() {
     ]);
   });
 
-  test('Should read the nested value', () async {
-    final providerA = FutureProvider(
-      (ref) => Future.delayed(
-        Duration(milliseconds: 50),
-        () => 'AAA',
-      ),
-      debugLabel: 'providerA',
-    );
-    final providerB = FutureProvider(
-      (ref) => Future.delayed(
-        Duration(milliseconds: 100),
-        () => 'BBB',
-      ),
-      debugLabel: 'providerB',
-    );
-    final providerC = FutureProvider((ref) async {
-      final a = await ref.future(providerA);
-      final b = await ref.future(providerB);
-      return '$a $b CCC';
-    }, debugLabel: 'providerC');
+  test('Should await other FutureProviders', () async {
     final observer = RefenaHistoryObserver.all();
     final ref = RefenaContainer(
       observers: [observer],
     );
 
-    expect(ref.read(providerC), AsyncValue<String>.loading());
-    expect(await ref.future(providerC), 'AAA BBB CCC');
+    final parentProviderA = FutureProvider(
+      (ref) => Future.delayed(Duration(milliseconds: 50), () => 'AAA'),
+    );
+
+    final parentProviderB = FutureProvider(
+      (ref) => Future.delayed(Duration(milliseconds: 100), () => 'BBB'),
+    );
+
+    final childProvider = FutureProvider(
+      (ref) async {
+        final a = await ref.future(parentProviderA);
+        final b = await ref.future(parentProviderB);
+        return '$a $b CCC';
+      },
+    );
+
+    expect(ref.read(childProvider), AsyncValue<String>.loading());
+
+    expect(await ref.future(childProvider), 'AAA BBB CCC');
     expect(
-      ref.read(providerC),
+      ref.read(childProvider),
       AsyncValue.data('AAA BBB CCC'),
     );
 
     // Check events
-    final notifierA =
+    final parentNotifierA =
         ref.anyNotifier<FutureProviderNotifier<String>, AsyncValue<String>>(
-      providerA,
+      parentProviderA,
     );
-    final notifierB =
+    final parentNotifierB =
         ref.anyNotifier<FutureProviderNotifier<String>, AsyncValue<String>>(
-      providerB,
+      parentProviderB,
     );
-    final notifierC =
+    final childNotifier =
         ref.anyNotifier<FutureProviderNotifier<String>, AsyncValue<String>>(
-      providerC,
+      childProvider,
     );
 
     // providerA and providerB are initialized immediately
     // providerC is initialized after the await of providerA
     expect(observer.history, [
       ProviderInitEvent(
-        provider: providerA,
-        notifier: notifierA,
+        provider: parentProviderA,
+        notifier: parentNotifierA,
         cause: ProviderInitCause.access,
         value: AsyncValue<String>.loading(),
       ),
       ProviderInitEvent(
-        provider: providerC,
-        notifier: notifierC,
+        provider: childProvider,
+        notifier: childNotifier,
         cause: ProviderInitCause.access,
         value: AsyncValue<String>.loading(),
       ),
       ChangeEvent(
-        notifier: notifierA,
+        notifier: parentNotifierA,
         action: null,
         prev: AsyncValue<String>.loading(),
         next: AsyncValue.data('AAA'),
         rebuild: [],
       ),
       ProviderInitEvent(
-        provider: providerB,
-        notifier: notifierB,
+        provider: parentProviderB,
+        notifier: parentNotifierB,
         cause: ProviderInitCause.access,
         value: AsyncValue<String>.loading(),
       ),
       ChangeEvent(
-        notifier: notifierB,
+        notifier: parentNotifierB,
         action: null,
         prev: AsyncValue<String>.loading(),
         next: AsyncValue.data('BBB'),
         rebuild: [],
       ),
       ChangeEvent(
-        notifier: notifierC,
+        notifier: childNotifier,
         action: null,
         prev: AsyncValue<String>.loading(),
         next: AsyncValue.data('AAA BBB CCC'),
+        rebuild: [],
+      ),
+    ]);
+  });
+
+  test('Should rebuild on parent change', () async {
+    final observer = RefenaHistoryObserver.all();
+    final ref = RefenaContainer(
+      observers: [observer],
+    );
+
+    final parentProvider = StateProvider((ref) => 0);
+    final childProvider = FutureProvider((ref) async {
+      final value = ref.watch(parentProvider);
+      await Future.delayed(Duration(milliseconds: 50));
+      return value * 2;
+    });
+
+    expect(ref.read(childProvider), AsyncValue<int>.loading());
+
+    expect(await ref.future(childProvider), 0);
+    expect(
+      ref.read(childProvider),
+      AsyncValue.data(0),
+    );
+
+    final parentNotifier = ref.anyNotifier<StateNotifier<int>, int>(
+      parentProvider,
+    );
+    final childNotifier =
+        ref.anyNotifier<FutureProviderNotifier<int>, AsyncValue<int>>(
+      childProvider,
+    );
+
+    // Check dependency graph
+    expect(parentNotifier.dependencies, isEmpty);
+    expect(parentNotifier.dependents, {childNotifier});
+    expect(childNotifier.dependencies, {parentNotifier});
+    expect(childNotifier.dependents, isEmpty);
+
+    // Change parent
+    ref.notifier(parentProvider).setState((old) => 10);
+    await skipAllMicrotasks();
+
+    expect(ref.read(childProvider), AsyncValue<int>.loading(0));
+
+    expect(await ref.future(childProvider), 20);
+    expect(
+      ref.read(childProvider),
+      AsyncValue.data(20),
+    );
+
+    // Check events
+    expect(observer.history, [
+      ProviderInitEvent(
+        provider: parentProvider,
+        notifier: parentNotifier,
+        cause: ProviderInitCause.access,
+        value: 0,
+      ),
+      ProviderInitEvent(
+        provider: childProvider,
+        notifier: childNotifier,
+        cause: ProviderInitCause.access,
+        value: AsyncValue<int>.loading(),
+      ),
+      ChangeEvent(
+        notifier: childNotifier,
+        action: null,
+        prev: AsyncValue<int>.loading(),
+        next: AsyncValue.data(0),
+        rebuild: [],
+      ),
+      ChangeEvent(
+        notifier: parentNotifier,
+        action: null,
+        prev: 0,
+        next: 10,
+        rebuild: [childNotifier],
+      ),
+      RebuildEvent(
+        rebuildable: childNotifier,
+        causes: [
+          ChangeEvent<int>(
+            notifier: parentNotifier,
+            action: null,
+            prev: 0,
+            next: 10,
+            rebuild: [childNotifier],
+          ),
+        ],
+        prev: AsyncValue.data(0),
+        next: AsyncValue<int>.loading(0),
+        rebuild: [],
+      ),
+      ChangeEvent(
+        notifier: childNotifier,
+        action: null,
+        prev: AsyncValue<int>.loading(0),
+        next: AsyncValue.data(20),
         rebuild: [],
       ),
     ]);