+++ refena_flutter/lib/src/scope.dart
@@ -221,8 +221,8 @@ class RefenaScope extends StatefulWidget implements RefenaContainer {
   }
 
   @override
-  Future<T> future<N extends AsyncNotifier<T>, T>(
-    AsyncNotifierProvider<N, T> provider,
+  Future<T> future<N extends BaseAsyncNotifier<T>, T>(
+    BaseProvider<N, AsyncValue<T>> provider,
   ) {
     return _container.future(provider);
   }